+++ data-final-fixes.lua
@@ -1,0 +1,11 @@
+data.raw["technology"]["fluid-handling"].effects =
+{
+  {
+    type = "unlock-recipe",
+    recipe = "storage-tank"
+  },
+  {
+    type = "unlock-recipe",
+    recipe = "solid-storage-tank"
+  },
+}