+++ data-updates.lua
@@ -24,8 +24,7 @@ function fw.is_item_convertable(item)
         return false
     end
 
-    --exclude ores
-    if fw.items_that_should_not_be_converted[item.name] == 1 then
+    if fw.items_that_should_not_be_converted[item.name] then
         return false
     end
 
@@ -36,9 +35,16 @@ function fw.is_recipe_convertable(item)
     return true
 end
 
-function fw.does_item_should_have_solid_variant(item)
+function fw.should_item_have_solid_variant(item)
     for _, solid in pairs(fw.prototypes_with_solid_variant) do
         if solid.name == item.name or solid.type == item.type then
+            if solid.except ~= nil then
+                for _, exc in pairs(solid.except) do
+                    if exc == item.name then
+                        return false
+                    end
+                end
+            end
             return true
         end
     end
@@ -54,11 +60,16 @@ function fw.fluidify_item(item)
     item.max_temperature = 100
     item.pressure_to_speed_ratio = 0.4
     item.flow_to_energy_ratio = 0.59
-    item.order = "a[fluid]-b[" .. item.name .. "]"
+    item.subgroup = "fluid-" .. item.subgroup
+    item.icons = {
+        {icon = item.icon},
+        {icon = "__fluidworld__/graphics/icons/fluid-overlay.png"}
+    }
+    item.icon = nil
 end
 
-function fw.fluidify_recipe(item)
-    local recipe = data.raw["recipe"][item.name]
+function fw.fluidify_recipe(name)
+    local recipe = data.raw["recipe"][name]
     -- if recipe == nil then
     --     print(item.name)
     -- end
@@ -69,12 +80,18 @@ function fw.fluidify_recipe(item)
             recipe.category = "fluid-" .. recipe.category
         end
 
+        local enabled = recipe.enabled
+        if enabled == nil and recipe.normal ~= nil then
+            enabled = recipe.normal.enabled
+        end
+        recipe.enabled = enabled
+
         local ingredients = recipe.ingredients
         if ingredients == nil then
             ingredients = recipe.normal.ingredients
             recipe.ingredients = recipe.normal.ingredients
         end
-        for _, ingredient in pairs(ingredients) do
+        for key, ingredient in pairs(ingredients) do
             if ingredient.type ~= "fluid" then
                 local ing_name, ing_amount
                 -- print(serpent.block(ingredient))
@@ -86,9 +103,9 @@ function fw.fluidify_recipe(item)
                     ing_amount = ingredient.amount
                 end
 
-                if fw.items_that_should_not_be_converted[ing_name] ~= 1 then                    
+                if fw.items_that_should_not_be_converted[ing_name] ~= true then
                     -- print(ing_name .. " - " .. serpent.block(ingredient))
-                    ingredient = {type = "fluid", name = ing_name, amount = ing_amount * 10}
+                    ingredients[key] = {type = "fluid", name = ing_name, amount = ing_amount * 10}
                 end
             end
         end
@@ -98,7 +115,7 @@ function fw.fluidify_recipe(item)
             results = recipe.normal.results
         end
         if results ~= nil then
-            for _, result in pairs(results) do
+            for key, result in pairs(results) do
                 if result.type ~= "fluid" then
                     local res_name, res_amount
                     if result.type == nil then
@@ -109,8 +126,15 @@ function fw.fluidify_recipe(item)
                         res_amount = result.amount
                     end
 
-                    if fw.items_that_should_not_be_converted[res_name] ~= 1 then
-                        result = {type = "fluid", name = res_name, amount = res_amount * 10}
+                    if res_amount == nil then
+                        results[key] = {
+                            type = "fluid",
+                            name = result.name,
+                            probability = result.probability,
+                            amount = result.amount * 10
+                        }
+                    elseif fw.items_that_should_not_be_converted[res_name] ~= true then
+                        results[key] = {type = "fluid", name = res_name, amount = res_amount * 10}
                     end
                 end
             end
@@ -127,7 +151,7 @@ function fw.fluidify_recipe(item)
                 if result_count == nil then
                     result_count = 1
                 end
-                
+
                 recipe.results = {{type = "fluid", name = res, amount = result_count * 10}}
                 recipe.result = nil
                 recipe.result_count = nil
@@ -136,46 +160,497 @@ function fw.fluidify_recipe(item)
             end
         end
 
-        -- if recipe.name == "steam-engine" then 
-        --     print("!!! - " .. serpent.block(recipe))
-        -- end
+        if recipe.name == "sulfuric-acid" then
+            print("!!! - " .. serpent.block(recipe))
+        end
     end
 end
 
-function fw.change_solid_item_recipe(item)
+function fw.add_solidify_recipe(item)
     local recipe = data.raw["recipe"][item.name]
     if recipe ~= nil then
-        recipe.category = "solidifying"
-        recipe.ingredients = {{type = "fluid", name = item.name, amount = 10}}
-        recipe.results = {{type = "item", name = item.name, amount = 1}}
-    end
-end
-
-function fw.log(item, item_name, message)
-    if item.name == item_name then
-        print(message)
+        local solid_recipe = util.table.deepcopy(recipe)
+        solid_recipe.name = "solid-" .. item.name
+        solid_recipe.category = "solidifying"
+        solid_recipe.ingredients = {{type = "fluid", name = item.name, amount = 10}}
+        solid_recipe.results = {{type = "item", name = item.name, amount = 1}}
+        data:extend({solid_recipe})
     end
 end
 
 function fw.convert_item(item)
     if fw.is_item_convertable(item) then
-        -- print(item.name)
-        local new_item = fw.deepcopy(item)
+        local new_item = util.table.deepcopy(item)
         fw.fluidify_item(new_item)
         data:extend({new_item})
 
-        if fw.does_item_should_have_solid_variant(item) then
-            fw.change_solid_item_recipe(item)
+        fw.fluidify_recipe(item.name)
+
+        if fw.should_item_have_solid_variant(item) then
+            fw.add_solidify_recipe(item)
         end
 
-        -- else
-        --     fw.log(item, "flamethrower-turret", "6")
-        --     fw.fluidify_item(item)
-        --     fw.log(item, "flamethrower-turret", "7")
-        -- end
-
-        fw.fluidify_recipe(item)
+    -- else
+    --     fw.log(item, "flamethrower-turret", "6")
+    --     fw.fluidify_item(item)
+    --     fw.log(item, "flamethrower-turret", "7")
+    -- end
     end
+end
+
+function fw.update_technologies()
+    for _, tech in pairs(data.raw["technology"]) do
+        if tech.effects ~= nil then
+            local new_effects = {}
+            local added = false
+            for _, effect in pairs(tech.effects) do
+                if effect.type == "unlock-recipe" and data.raw["recipe"]["solid-" .. effect.recipe] ~= nil then
+                    new_effects[#new_effects + 1] = {
+                        type = "unlock-recipe",
+                        recipe = "solid-" .. effect.recipe
+                    }
+                    added = true
+                end
+            end
+            if added then
+                for _, new in pairs(new_effects) do
+                    tech.effects[#tech.effects + 1] = new
+                end
+            end
+        end
+    end
+end
+
+function fw.update_assemblers()
+    local am1 = data.raw["assembling-machine"]["assembling-machine-1"]
+    am1.crafting_categories = {"fluid-crafting"}
+    am1.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-1, -2}}}
+        },
+        {
+            production_type = "output",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {0, 2}}}
+        }
+    }
+
+    local am2 = data.raw["assembling-machine"]["assembling-machine-2"]
+    am2.crafting_categories = {"fluid-crafting", "fluid-advanced-crafting", "fluid-crafting-with-fluid"}
+    am2.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {2, 0}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-2, 0}}}
+        },
+        {
+            production_type = "output",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {0, 2}}}
+        }
+    }
+
+    local am3 = data.raw["assembling-machine"]["assembling-machine-3"]
+    am3.crafting_categories = {"fluid-crafting", "fluid-advanced-crafting", "fluid-crafting-with-fluid"}
+    am3.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {2, 1}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-2, 1}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {2, 0}}}
+        },
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-2, 0}}}
+        },
+        {
+            production_type = "output",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {0, 2}}}
+        }
+    }
+end
+
+function fw.update_buildings()
+    local cp = data.raw["assembling-machine"]["chemical-plant"]
+    cp.crafting_categories = {"fluid-chemistry"}
+    cp.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_level = 1,
+            pipe_connections = {{type = "input", position = {-1, 2}}}
+        },
+        {
+            production_type = "output",
+            pipe_covers = pipecoverspictures(),
+            base_level = 1,
+            pipe_connections = {{position = {1, 2}}}
+        }
+    }
+
+    data.raw["assembling-machine"]["oil-refinery"].crafting_categories = {"fluid-oil-processing"}
+
+    local cf = data.raw["assembling-machine"]["centrifuge"]
+    cf.crafting_categories = {"fluid-centrifuging"}
+    cf.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {-1, -2}}}
+        },
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {1, -2}}}
+        },
+        {
+            production_type = "output",
+            pipe_covers = pipecoverspictures(),
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {-1, 2}}}
+        },
+        {
+            production_type = "output",
+            pipe_covers = pipecoverspictures(),
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {1, 2}}}
+        }
+    }
+end
+
+function fw.update_furnaces()
+    local sf = data.raw["furnace"]["stone-furnace"]
+    sf.crafting_categories = {"fluid-smelting"}
+    sf.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {0.5, 1.5}}}
+        },
+        {
+            production_type = "output",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {-0.5, -1.5}}}
+        }
+    }
+
+    local steelf = data.raw["furnace"]["steel-furnace"]
+    steelf.crafting_categories = {"fluid-smelting"}
+    steelf.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {0.5, 1.5}}}
+        },
+        {
+            production_type = "output",
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {-0.5, -1.5}}}
+        }
+    }
+
+    local elf = data.raw["furnace"]["electric-furnace"]
+    elf.crafting_categories = {"fluid-smelting"}
+    elf.fluid_boxes = {
+        {
+            production_type = "input",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = -1,
+            pipe_connections = {{type = "input", position = {1, 2}}}
+        },
+        {
+            production_type = "output",
+            pipe_picture = assembler2pipepictures(),
+            pipe_covers = pipecoverspictures(),
+            base_area = 10,
+            base_level = 1,
+            pipe_connections = {{type = "output", position = {-1, -2}}}
+        }
+    }
+end
+
+function fw.add_solidifier()
+    data:extend(
+        {
+            {
+                type = "assembling-machine",
+                name = "solidifier",
+                icon = "__fluidworld__/graphics/icons/solidifier.png",
+                icon_size = 32,
+                flags = {"placeable-neutral", "placeable-player", "player-creation"},
+                minable = {hardness = 0.2, mining_time = 0.5, result = "solidifier"},
+                max_health = 400,
+                corpse = "big-remnants",
+                dying_explosion = "medium-explosion",
+                resistances = {
+                    {
+                        type = "fire",
+                        percent = 70
+                    }
+                },
+                fluid_boxes = {
+                    {
+                        production_type = "input",
+                        pipe_picture = assembler3pipepictures(),
+                        pipe_covers = pipecoverspictures(),
+                        base_area = 10,
+                        base_level = -1,
+                        pipe_connections = {{type = "input", position = {0, -2}}},
+                        secondary_draw_orders = {north = -1}
+                    },
+                    {
+                        production_type = "output",
+                        pipe_picture = assembler3pipepictures(),
+                        pipe_covers = pipecoverspictures(),
+                        base_area = 10,
+                        base_level = -1,
+                        pipe_connections = {{type = "output", position = {0, 2}}}
+                    },
+                    off_when_no_fluid_recipe = false
+                },
+                open_sound = {filename = "__base__/sound/machine-open.ogg", volume = 0.85},
+                close_sound = {filename = "__base__/sound/machine-close.ogg", volume = 0.75},
+                vehicle_impact_sound = {filename = "__base__/sound/car-metal-impact.ogg", volume = 0.65},
+                working_sound = {
+                    sound = {
+                        {
+                            filename = "__base__/sound/assembling-machine-t3-1.ogg",
+                            volume = 0.8
+                        },
+                        {
+                            filename = "__base__/sound/assembling-machine-t3-2.ogg",
+                            volume = 0.8
+                        }
+                    },
+                    idle_sound = {filename = "__base__/sound/idle1.ogg", volume = 0.6},
+                    apparent_volume = 1.5
+                },
+                collision_box = {{-1.2, -1.2}, {1.2, 1.2}},
+                selection_box = {{-1.5, -1.5}, {1.5, 1.5}},
+                fast_replaceable_group = "assembling-machine",
+                animation = {
+                    layers = {
+                        {
+                            filename = "__fluidworld__/graphics/entity/solidifier/solidifier.png",
+                            priority = "high",
+                            width = 108,
+                            height = 119,
+                            frame_count = 32,
+                            line_length = 8,
+                            shift = util.by_pixel(0, -0.5),
+                            hr_version = {
+                                filename = "__fluidworld__/graphics/entity/solidifier/hr-solidifier.png",
+                                priority = "high",
+                                width = 214,
+                                height = 237,
+                                frame_count = 32,
+                                line_length = 8,
+                                shift = util.by_pixel(0, -0.75),
+                                scale = 0.5
+                            }
+                        },
+                        {
+                            filename = "__fluidworld__/graphics/entity/solidifier/solidifier-shadow.png",
+                            priority = "high",
+                            width = 130,
+                            height = 82,
+                            frame_count = 32,
+                            line_length = 8,
+                            draw_as_shadow = true,
+                            shift = util.by_pixel(28, 4),
+                            hr_version = {
+                                filename = "__fluidworld__/graphics/entity/solidifier/hr-solidifier-shadow.png",
+                                priority = "high",
+                                width = 260,
+                                height = 162,
+                                frame_count = 32,
+                                line_length = 8,
+                                draw_as_shadow = true,
+                                shift = util.by_pixel(28, 4),
+                                scale = 0.5
+                            }
+                        }
+                    }
+                },
+                crafting_categories = {"solidifying"},
+                crafting_speed = 1.25,
+                energy_source = {
+                    type = "electric",
+                    usage_priority = "secondary-input",
+                    emissions = 0.03 / 3.5
+                },
+                energy_usage = "210kW",
+                ingredient_count = 6,
+                module_specification = {
+                    module_slots = 4
+                },
+                allowed_effects = {"consumption", "speed", "productivity", "pollution"}
+            },
+            {
+                type = "recipe",
+                name = "solidifier",
+                category = "fluid-crafting",
+                enabled = true,
+                ingredients = {
+                    {type = "fluid", name = "electronic-circuit", amount = 50},
+                    {type = "fluid", name = "iron-gear-wheel", amount = 50}
+                },
+                result = "solidifier"
+            },
+            {
+                type = "item",
+                name = "solidifier",
+                icon = "__fluidworld__/graphics/icons/solidifier.png",
+                icon_size = 32,
+                flags = {"goes-to-quickbar"},
+                subgroup = "solidifier",
+                order = "a[solidifier]",
+                place_result = "solidifier",
+                stack_size = 50
+            }
+        }
+    )
+end
+
+function fw.fine_tune()
+    data:extend(
+        {
+            {
+                type = "recipe",
+                name = "liquify-used-up-uranium-fuel-cell",
+                ingredients = {
+                    {"used-up-uranium-fuel-cell", 1}
+                },
+                results = {
+                    {type = "fluid", name = "used-up-uranium-fuel-cell", amount = 10}
+                },
+                category = "solidifying",
+                subgroup = "intermediate-product",
+                order = "r[uranium-processing]-a[used-up-uranium-fuel-cell]"
+            }
+        }
+    )
+    local ef = data.raw["technology"]["nuclear-fuel-reprocessing"].effects
+    ef[#ef + 1] = {
+        type = "unlock-recipe",
+        recipe = "liquify-used-up-uranium-fuel-cell"
+    }
 end
 
 function fw.start_converting()
@@ -184,98 +659,20 @@ function fw.start_converting()
             fw.convert_item(item)
         end
     end
+
+    for _, rec in pairs(fw.additional_recipies_to_convert) do
+        fw.fluidify_recipe(rec)
+    end
+
+    fw.update_technologies()
+
+    fw.update_assemblers()
+    fw.update_buildings()
+    fw.update_furnaces()
+
+    fw.add_solidifier()
+
+    fw.fine_tune()
 end
 
 fw.start_converting()
--- error("STOP")
-
--- for _, recipe in pairs(data.raw["recipe"]) do
---     if fw.is_recipe_convertable(item) then
---         local new_fluid_recipe = {
---         }
---         data:extend({new_fluid_recipe})
---     end
--- end
-
--- local x = {}
--- local y = deepcopy(x)
-
--- data:extend(
---     {
---         {
---             type = "recipe",
---             name = "iron-chest",
---             category = "fluid-crafting",
---             --enabled = false,
---             --energy_required = 5,
---             ingredients = {
---                 {type = "fluid", name = "iron-plate", amount = 8}
---             },
---             results = {
---                 {type = "fluid", name = "iron-chest", amount = 1}
---             },
---             --icon = "__base__/graphics/icons/fluid/basic-oil-processing.png",
---             --subgroup = "fluid-recipes",
---             order = "a[oil-processing]-a[iron-chest]"
---         },
---         {
---             type = "recipe",
---             name = "iron-plate",
---             category = "fluid-smelting",
---             energy_required = 3.5,
---             ingredients = {{"iron-ore", 1}},
---             results = {{type = "fluid", name = "iron-plate", amount = 1}}
---         }
---     }
--- )
-
--- local new_stone_furnace = data.raw["furnace"]["stone-furnace"]
--- new_stone_furnace.collision_box = {{-0.7, -0.7}, {0.7, 0.7}}
--- new_stone_furnace.selection_box = {{-1, -1}, {1, 1}}
--- new_stone_furnace.crafting_categories = {"fluid-smelting"}
--- new_stone_furnace.fluid_boxes = {
---     {
---         production_type = "output",
---         base_area = 1,
---         base_level = 1,
---         pipe_covers = pipecoverspictures(),
---         pipe_connections = {{position = {-0.5, -1.5}}}
---     }
--- }
-
--- local new_burner_mining_drill = data.raw["mining-drill"]["burner-mining-drill"]
--- new_burner_mining_drill.resource_categories = {"basic-fluid"}
--- new_burner_mining_drill.collision_box = {{-1.1, -1.1}, {1.1, 1.1}}
--- new_burner_mining_drill.selection_box = {{-1.2, -1.2}, {1.2, 1.2}}
--- new_burner_mining_drill.vector_to_place_result = {0, 0}
--- new_burner_mining_drill.output_fluid_box = {
---     production_type = "output",
---     base_area = 1,
---     base_level = 1,
---     pipe_covers = pipecoverspictures(),
---     pipe_connections = {{position = {-1, -2}}}
--- }
--- new_burner_mining_drill.input_fluid_box = {
---     production_type = "input-output",
---     base_area = 10,
---     height = 2,
---     base_level = -1,
---     pipe_picture = assembler2pipepictures(),
---     pipe_covers = pipecoverspictures(),
---     pipe_connections = {{ position = {-1, 2}}}
--- }
-
--- local new_coal = data.raw["resource"]["coal"]
--- new_coal.mining_particle = nil
--- new_coal.category = "basic-fluid"
--- new_coal.minable.result = nil
--- new_coal.minable.results = {
---     {
---         type = "fluid",
---         name = "coal",
---         amount_min = 10,
---         amount_max = 10,
---         probability = 1
---     }
--- }
--- new_coal.minable.required_fluid = "coal"