+++ prototypes/categories.lua
@@ -9,7 +9,21 @@ function fw.add_categories()
     data:extend(new_categories)
 end
 
+function fw.add_subgroups()
+    local new_subgroups = {}
+    for _, sg in pairs(data.raw["item-subgroup"]) do
+        new_subgroups[#new_subgroups + 1] = {
+            type = "item-subgroup",
+            name = "fluid-" .. sg.name,
+            group = sg.group,
+            order = sg.order .. "a"
+        }
+    end
+    data:extend(new_subgroups)
+end
+
 fw.add_categories()
+fw.add_subgroups()
 
 data:extend(
     {
@@ -17,7 +31,13 @@ data:extend(
         {
             type = "recipe-category",
             name = "solidifying"
-        }
+        },
         --SUBGROUPS
+        {
+            type = "item-subgroup",
+            name = "solidifier",
+            group = "production",
+            order = "da"
+        }
     }
 )