+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0 (4 October 2018)
+
+* Added "Solidifier" entity that is making items from fluids for player to be able to use
+* Added solid recipes
+* Added pipe connectors to all supported buildings
+* Updated fluid recipe icons and descriptions
+
 ## 0.1.0 (20 September 2018)
 
 * Implemented items and recipes automatic conversion